+++ src/lib/components/Feedback.svelte
@@ -19,10 +19,11 @@
 	span {
 		border-left: calc(var(--thick-border-width) * 2) solid var(--salient-color);
 		background: var(--salient-color-faded);
-		padding: calc(var(--spacing) / 3) calc(var(--spacing));
+		padding: calc(var(--spacing)) calc(var(--spacing));
 		font-size: var(--small-font-size);
 		border-top-right-radius: var(--roundedness);
 		border-bottom-right-radius: var(--roundedness);
+		border-radius: var(--roundedness);
 	}
 
 	p {