+++ src/lib/components/Header.svelte
@@ -56,7 +56,7 @@
 		display: flex;
 		flex-direction: row;
 		flex-wrap: wrap;
-		gap: calc(2 * var(--spacing));
+		gap: var(--spacing);
 		row-gap: var(--spacing);
 		align-items: center;
 		padding-bottom: var(--spacing);